fix CreatePoolGyroECLP token sorting; move derived params calc into buildCall()
Sorting tokens must also adjust the ECLP params since otherwise, the
meaning of the params changes just because of the token addresses. This
is required here b/c the ECLP is not (necessarily) symmetric.

The derived params calculation is affected by this, so I moved it into
the same call and removed it from CreatePoolGyroECLPInput. Also adjusted
tests.

NB the same adjustment will be required for the 2CLP's price bounds.