fix CreatePoolGyroECLP token sorting; move derived params calc into b… (#685)
* fix CreatePoolGyroECLP token sorting; move derived params calc into buildCall()

Sorting tokens must also adjust the ECLP params since otherwise, the
meaning of the params changes just because of the token addresses. This
is required here b/c the ECLP is not (necessarily) symmetric.

The derived params calculation is affected by this, so I moved it into
the same call and removed it from CreatePoolGyroECLPInput. Also adjusted
tests.

NB the same adjustment will be required for the 2CLP's price bounds.

* format

* add create gyro integration test for tokens out of order

* clean up and update eclp params for integration test

* fix type any

* test inverted eclp params input

* improve readability

* scale to 100 decimals for reciprical calculation

* make derived eclp params optional

* integration test for inverted params with relative tolerance

* unit tests for gyro creation helper that normalizes prams

* add changeset

---------

Co-authored-by: Matt Pereira <mattpereira.dev@gmail.com>